requirements.txt
django==1.11
zappa
pillow
wand
boto3
python-memcached
django-cors-headers
django-redis
sorl-thumbnail
django-storages
coverage